api/requirements.txt
fastapi
uvicorn
openai
requests
langchain
pydantic
langchain_ollama
langchain_community
langchain_qdrant
langchainhub
python-docx
python-multipart